Finish RPS project v1 + create script_v2.js
Finished the RPS game without refactoring it;

Added a new file script_v2.js that contains the code from script.js
without any comments, so refactoring is a bit more convenient for this
type of project. Reasoning: Comments do take way to much space, so
making changes is kind of a pain.